+++ haiku_rag_slim/haiku/rag/ingester/pollers/fs.py
@@ -11,8 +11,8 @@ from haiku.rag.ingester.sources.filter import FileFilter
 from haiku.rag.telemetry import logfire
 
 if TYPE_CHECKING:
+    from haiku.rag.circuit_breaker import CircuitBreaker
     from haiku.rag.config import FSSourceConfig
-    from haiku.rag.ingester.pollers.circuit_breaker import CircuitBreaker
     from haiku.rag.ingester.sources.fs import FSSource
 
 logger = logging.getLogger(__name__)
